feat: `SapTableBody` and `SapTableRow` as separate struct
