+++ src/webdynpro/element/complex/sap_table/mod.rs
@@ -116,3 +116,4 @@ impl<'a> SapTable<'a> {
 pub mod body;
 pub mod cell;
 pub mod property;
+pub mod row;
